@@ -14,9 +14,7 @@ import com.rollbar.android.Rollbar;
 
 public class MyApplication extends Application {
     private static MyApplication instance;
-    public static  MyApplication getInstance() {
-        return instance;
-    }
+    boolean calendarIsOpen;
     @Override
     public void onCreate() {
         super.onCreate();
@@ -31,6 +29,22 @@ public class MyApplication extends Application {
             }
         });
 
+    }
+
+    public static synchronized MyApplication getInstance(){
+        return instance == null ? new MyApplication() : instance;
+    }
+
+    public static void setInstance(MyApplication instance) {
+        MyApplication.instance = instance;
+    }
+
+    public boolean isCalendarIsOpen() {
+        return calendarIsOpen;
+    }
+
+    public void setCalendarIsOpen(boolean calendarIsOpen) {
+        this.calendarIsOpen = calendarIsOpen;
     }
 
     /*This methods belongs to enable Multidexing for Application and Fixing crashing issue for lollipop and below it.*/
